Add CR comments #1
