+++ src/components/ContactDetails/ContactDetails.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 const ContactDetails: React.FC = () => {
-  let { id } = useParams();
+  const { id: newbieId } = useParams();
 
   return (
     <>
       <h1>Contact Details</h1>
-      <p>ID: {id}</p>
+      <p>ID: {newbieId}</p>
     </>
   );
 };
