+++ src/components/TaskDetails/TaskDetails.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 const TaskDetails: React.FC = () => {
-  let { taskId } = useParams();
+  const { taskId } = useParams();
 
   return (
     <>
